+++ cypress/cypress/component/Datepicker/DatePickerInput.cy.tsx
@@ -1,14 +1,17 @@
 /* eslint-disable max-lines */
-import { DatePickerInput, Form } from "react-hook-form-components";
+import { DatePickerInput, Form, setUtcTimeToZero } from "react-hook-form-components";
 import "react-datepicker/dist/react-datepicker.css";
 import { faker } from "@faker-js/faker";
 import * as yup from "yup";
 import { yupResolver } from "@hookform/resolvers/yup";
 
-it("basic example works", () => {
+it("selecting today works", () => {
   const name = faker.random.alpha(10);
+  const todayMidnight = new Date();
+  setUtcTimeToZero(todayMidnight);
+
   const schema = yup.object().shape({
-    [name]: yup.string().required(),
+    [name]: yup.date().required(),
   });
 
   cy.mount(
@@ -22,18 +25,18 @@ it("basic example works", () => {
   cy.contains("label", name).click();
   cy.get(".react-datepicker__day--today").click();
   cy.get("input[type=submit]").click({ force: true });
-
-  const todayMidnight = new Date();
-  todayMidnight.setHours(0, 0, 0, 0);
-  todayMidnight.setMinutes(todayMidnight.getMinutes() - todayMidnight.getTimezoneOffset());
-
-  cy.get("@onSubmitSpy").should("be.calledOnceWith", { [name]: todayMidnight.toString() });
+  cy.get("@onSubmitSpy")
+    .its("lastCall.args.0")
+    .should("deep.equal", { [name]: todayMidnight });
 });
 
-it.only("basic example works", () => {
+it("setting intial value as iso string works", () => {
   const name = faker.random.alpha(10);
+  const randomDate = faker.date.future();
+  setUtcTimeToZero(randomDate);
+
   const schema = yup.object().shape({
-    [name]: yup.string().required(),
+    [name]: yup.date().required(),
   });
 
   cy.mount(
@@ -41,7 +44,7 @@ it.only("basic example works", () => {
       onSubmit={cy.spy().as("onSubmitSpy")}
       resolver={yupResolver(schema)}
       defaultValues={{
-        [name]: "2023-01-27T00:00:00.000Z",
+        [name]: randomDate.toISOString(),
       }}
     >
       <DatePickerInput name={name} label={name} />
@@ -51,12 +54,38 @@ it.only("basic example works", () => {
   );
 
   cy.get("input[type=submit]").click({ force: true });
+  cy.get("@onSubmitSpy")
+    .its("lastCall.args.0")
+    .should("deep.equal", { [name]: randomDate });
+});
 
-  const todayMidnight = new Date(2023, 0, 27);
-  todayMidnight.setHours(0, 0, 0, 0);
-  todayMidnight.setMinutes(todayMidnight.getMinutes() - todayMidnight.getTimezoneOffset());
+it("setting intial value as date object works", () => {
+  const name = faker.random.alpha(10);
+  const randomDate = faker.date.future();
+  setUtcTimeToZero(randomDate);
 
-  cy.get("@onSubmitSpy").should("be.calledOnceWith", { [name]: todayMidnight.toString() });
+  const schema = yup.object().shape({
+    [name]: yup.date().required(),
+  });
+
+  cy.mount(
+    <Form
+      onSubmit={cy.spy().as("onSubmitSpy")}
+      resolver={yupResolver(schema)}
+      defaultValues={{
+        [name]: randomDate,
+      }}
+    >
+      <DatePickerInput name={name} label={name} />
+
+      <input type={"submit"} />
+    </Form>,
+  );
+
+  cy.get("input[type=submit]").click({ force: true });
+  cy.get("@onSubmitSpy")
+    .its("lastCall.args.0")
+    .should("deep.equal", { [name]: randomDate });
 });
 
 it("is disabled", () => {