+++ src/index.ts
@@ -6,3 +6,4 @@ export * from "./lib/AsyncTypeaheadInput";
 export * from "./lib/types/Typeahead";
 export * from "./lib/types/LabelValueOption";
 export * from "./lib/DatePickerInput";
+export * from "./lib/helpers/dateUtils";