+++ src/lib/DatePickerInput.tsx
@@ -3,6 +3,8 @@ import { useSafeNameId } from "src/lib/hooks/useSafeNameId";
 import { FormGroupLayout } from "./FormGroupLayout";
 import { CommonInputProps } from "./types/CommonInputProps";
 import DatePicker, { ReactDatePickerProps } from "react-datepicker";
+import { useState } from "react";
+import { setUtcTimeToZero } from "./helpers/dateUtils";
 
 interface DatePickerInputProps<T extends FieldValues> extends Omit<CommonInputProps<T>, "onChange"> {
   datePickerProps?: Omit<ReactDatePickerProps, "onChange" | "selected" | "id" | "className" | "onBlur">;
@@ -13,63 +15,60 @@ const DatePickerInput = <T extends FieldValues>(props: DatePickerInputProps<T>) 
   const { disabled, label, helpText, datePickerProps, labelToolTip } = props;
   const { name, id } = useSafeNameId(props.name, props.id);
 
-  const { control } = useFormContext();
+  const { control, getValues } = useFormContext();
 
   const dateFormat = datePickerProps?.dateFormat || "dd.MM.yyyy";
   const calendarStartDay = datePickerProps?.calendarStartDay || 1;
+
+  const initialDate = getValues(name) as Date | null;
+  setUtcTimeToZero(initialDate);
+
+  const [selectedDate, setSelectedDate] = useState<Date | null>(initialDate);
 
   return (
     <Controller
       control={control}
       name={name}
-      render={({ field, fieldState: { error } }) => {
-        // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-        const selected = field.value;
+      render={({ field, fieldState: { error } }) => (
+        <FormGroupLayout helpText={helpText} name={name} id={id} label={label} labelToolTip={labelToolTip}>
+          <DatePicker
+            {...datePickerProps}
+            {...field}
+            id={id}
+            disabled={disabled}
+            className="form-control"
+            dateFormat={dateFormat}
+            calendarStartDay={calendarStartDay}
+            wrapperClassName={error ? "is-invalid" : ""}
+            selected={selectedDate}
+            ref={(elem) => {
+              // https://github.com/react-hook-form/react-hook-form/discussions/5413
+              // https://codesandbox.io/s/react-hook-form-focus-forked-yyhsi?file=/src/index.js
+              // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-explicit-any
+              elem && field.ref((elem as any).input);
+            }}
+            onBlur={(e) => {
+              if (props.onBlur) props.onBlur(e);
+              field.onBlur();
+            }}
+            onChange={(date) => {
+              setSelectedDate(date);
 
-        if (selected && selected instanceof Date) {
-          // reverting the changes that we are doing in the onChange handler
-          // otherwise the time will keep shifting by the timezone offset
-          selected.setMinutes(selected.getMinutes() + selected.getTimezoneOffset());
-        }
+              const utcTimeZeroDate = date == null ? null : new Date(date.getTime());
 
-        return (
-          <FormGroupLayout helpText={helpText} name={name} id={id} label={label} labelToolTip={labelToolTip}>
-            <DatePicker
-              {...datePickerProps}
-              {...field}
-              // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-              value={selected}
-              id={id}
-              disabled={disabled}
-              className="form-control"
-              dateFormat={dateFormat}
-              calendarStartDay={calendarStartDay}
-              wrapperClassName={error ? "is-invalid" : ""}
-              // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-              selected={selected}
-              ref={(elem) => {
-                // https://github.com/react-hook-form/react-hook-form/discussions/5413
-                // https://codesandbox.io/s/react-hook-form-focus-forked-yyhsi?file=/src/index.js
-                // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-explicit-any
-                elem && field.ref((elem as any).input);
-              }}
-              onBlur={(e) => {
-                if (props.onBlur) props.onBlur(e);
-                field.onBlur();
-              }}
-              onChange={(date) => {
-                // adding the timezone offset to the date so it's exactly 00:00:00 when converting to UTC
-                // (which JSON.stringify does)
-                if (date) date.setMinutes(date.getMinutes() - date.getTimezoneOffset());
+              // we set the time to utc 0 to avoid timezone issues, so it will be 0Z
+              // when JSON stringified
+              if (utcTimeZeroDate) {
+                setUtcTimeToZero(initialDate);
+              }
 
-                if (props.onChange) props.onChange(date);
+              if (props.onChange) props.onChange(utcTimeZeroDate);
 
-                field.onChange(date);
-              }}
-            />
-          </FormGroupLayout>
-        );
-      }}
+              field.onChange(utcTimeZeroDate);
+            }}
+          />
+        </FormGroupLayout>
+      )}
     />
   );
 };