+++ src/lib/Form.tsx
@@ -1,5 +1,6 @@
 import { ReactNode } from "react";
 import { DeepPartial, FieldValues, FormProvider, Resolver, SubmitHandler, useForm, UseFormReturn } from "react-hook-form";
+import { jsonIsoDateReviver } from "./helpers/dateUtils";
 
 interface FormProps<T extends FieldValues> {
   onSubmit: SubmitHandler<T>;
@@ -8,30 +9,10 @@ interface FormProps<T extends FieldValues> {
   children: ((formMethods: UseFormReturn<T, unknown>) => ReactNode) | ReactNode;
 }
 
-const jsonParseReviver = (_key: string, value: unknown) => {
-  console.log("reviving", value);
-  console.log("reviving", typeof value);
-  console.log("reviving", /\d{4}-[01]\d-[0-3]\dT[0-2]\d:[0-5]\d:[0-5]\d(([+-][0-2]\d:[0-5]\d)|Z)/.exec(value as string));
-
-  if (
-    typeof value == "string" &&
-    /(\d{4}-[01]\d-[0-3]\dT[0-2]\d:[0-5]\d:[0-5]\d\.\d+([+-][0-2]\d:[0-5]\d|Z))|(\d{4}-[01]\d-[0-3]\dT[0-2]\d:[0-5]\d:[0-5]\d([+-][0-2]\d:[0-5]\d|Z))|(\d{4}-[01]\d-[0-3]\dT[0-2]\d:[0-5]\d([+-][0-2]\d:[0-5]\d|Z))/.exec(
-      value,
-    )
-  ) {
-    console.log("reviving2", value);
-    return new Date(`${value}`);
-  }
-
-  return value;
-};
-
 const Form = <T extends FieldValues>({ children, onSubmit, resolver, defaultValues }: FormProps<T>) => {
   const revivedDefaultValues = defaultValues
-    ? (JSON.parse(JSON.stringify(defaultValues), jsonParseReviver) as DeepPartial<T>)
+    ? (JSON.parse(JSON.stringify(defaultValues), jsonIsoDateReviver) as DeepPartial<T>)
     : defaultValues;
-
-  console.log("revivedDefaultValues", revivedDefaultValues);
 
   const formMethods = useForm<T>({ resolver, defaultValues: revivedDefaultValues });
   const { handleSubmit } = formMethods;